create mock login system and install Floating Vue to create logout dropdown menu
This has the foundation of a login system and is ready to be hooked up to the events of whatever authentication system we use.

I added the Floating Vue library in order to easily handle the dropdown menu that houses the logout button. Other devs will need to run `npm install` in order to get this library.

With this commit, users should be able to click "login" to see the internal pages, and then click Profile icon > Log out. This is all mock behavior at the moment, and none of it authenticates the user or ties them to their personal data.