create start of user flow from login to dashboard
The login page used to lead to a test page with a bunch of links. Now it leads to the home page, which then links to the dashboard. The dashboard is blank at the moment, but the next step will be to fill it in so the user can access lessons.