create small starter set of color variables to standardize styles
This commit creates a starter set of color variables that can be re-used across the app and that will help standardize the styles. The color variables reside at the top of the `<style>` section of `App.vue`.

To use them, pass the name of the color variable to `var()`. For example, if you want to use our standard green (named `--var-color-green`) to set a background color, you'd use `background-color: var(--var-color-green);`