add environment file for api server url configuration
This change adds 2 environment files: `.env` and `.env.development`,
which contain a variable to specify the url of the api server. To begin
with this variable is set to `http://ugly.photography` for a production
mode build, and `http://localhost:8000` for development mode. The source
files have been updated to use the environment variable at build time.

The npm targets have also been updated to add a `build-dev` target, and
to place built files into subdirectories of `dist/` for each mode. The
readme file is also updated.

This change is being proposed to make configuring the api server
endpoint easier during the build process. Keeping the original value for
production builds will introduce the least surprise for now, while
having separate environment files will make it easy to change the values
in the future.